3-way split
