+++ amads/time/meter/examples.py
@@ -1,0 +1,1014 @@
+"""
+Examples of the start hierarchy structures.
+For reference and testing.
+Down to 1/32 note level [0, 0.125 ... ] in each case
+"""
+
+start_hierarchy_examples = {
+    "2/2": [
+        [0.0, 4.0],
+        [0.0, 2.0, 4.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+        ],
+    ],
+    "3/2": [
+        [0.0, 6.0],
+        [0.0, 2.0, 4.0, 6.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+        ],
+    ],
+    "4/2": [
+        [0.0, 8.0],
+        [0.0, 4.0, 8.0],
+        [0.0, 2.0, 4.0, 6.0, 8.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0, 8.0],
+        [
+            0.0,
+            0.5,
+            1.0,
+            1.5,
+            2.0,
+            2.5,
+            3.0,
+            3.5,
+            4.0,
+            4.5,
+            5.0,
+            5.5,
+            6.0,
+            6.5,
+            7.0,
+            7.5,
+            8.0,
+        ],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+            6.25,
+            6.5,
+            6.75,
+            7.0,
+            7.25,
+            7.5,
+            7.75,
+            8.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+            6.125,
+            6.25,
+            6.375,
+            6.5,
+            6.625,
+            6.75,
+            6.875,
+            7.0,
+            7.125,
+            7.25,
+            7.375,
+            7.5,
+            7.625,
+            7.75,
+            7.875,
+            8.0,
+        ],
+    ],
+    "2/4": [
+        [0.0, 2.0],
+        [0.0, 1.0, 2.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0],
+        [0.0, 0.25, 0.5, 0.75, 1.0, 1.25, 1.5, 1.75, 2.0],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+        ],
+    ],
+    "3/4": [
+        [0.0, 3.0],
+        [0.0, 1.0, 2.0, 3.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0],
+        [0.0, 0.25, 0.5, 0.75, 1.0, 1.25, 1.5, 1.75, 2.0, 2.25, 2.5, 2.75, 3.0],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+        ],
+    ],
+    "4/4": [
+        [0.0, 4.0],
+        [0.0, 2.0, 4.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+        ],
+    ],
+    "6/8": [
+        [0.0, 3.0],
+        [0.0, 1.5, 3.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0],
+        [0.0, 0.25, 0.5, 0.75, 1.0, 1.25, 1.5, 1.75, 2.0, 2.25, 2.5, 2.75, 3.0],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+        ],
+    ],
+    "9/8": [
+        [0.0, 4.5],
+        [0.0, 1.5, 3.0, 4.5],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+        ],
+    ],
+    "12/8": [
+        [0.0, 6.0],
+        [0.0, 3.0, 6.0],
+        [0.0, 1.5, 3.0, 4.5, 6.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+        ],
+    ],
+    "2+3/4": [
+        [0.0, 5.0],
+        [0.0, 2.0, 5.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+        ],
+    ],
+    "3+2/4": [
+        [0.0, 5.0],
+        [0.0, 3.0, 5.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+        ],
+    ],
+    "2+2+3/4": [
+        [0.0, 7.0],
+        [0.0, 2.0, 4.0, 7.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+            6.25,
+            6.5,
+            6.75,
+            7.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+            6.125,
+            6.25,
+            6.375,
+            6.5,
+            6.625,
+            6.75,
+            6.875,
+            7.0,
+        ],
+    ],
+    "3+2+2/4": [
+        [0.0, 7.0],
+        [0.0, 3.0, 5.0, 7.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+            6.25,
+            6.5,
+            6.75,
+            7.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+            6.125,
+            6.25,
+            6.375,
+            6.5,
+            6.625,
+            6.75,
+            6.875,
+            7.0,
+        ],
+    ],
+    "2+3+2/4": [
+        [0.0, 7.0],
+        [0.0, 2.0, 5.0, 7.0],
+        [0.0, 1.0, 2.0, 3.0, 4.0, 5.0, 6.0, 7.0],
+        [0.0, 0.5, 1.0, 1.5, 2.0, 2.5, 3.0, 3.5, 4.0, 4.5, 5.0, 5.5, 6.0, 6.5, 7.0],
+        [
+            0.0,
+            0.25,
+            0.5,
+            0.75,
+            1.0,
+            1.25,
+            1.5,
+            1.75,
+            2.0,
+            2.25,
+            2.5,
+            2.75,
+            3.0,
+            3.25,
+            3.5,
+            3.75,
+            4.0,
+            4.25,
+            4.5,
+            4.75,
+            5.0,
+            5.25,
+            5.5,
+            5.75,
+            6.0,
+            6.25,
+            6.5,
+            6.75,
+            7.0,
+        ],
+        [
+            0.0,
+            0.125,
+            0.25,
+            0.375,
+            0.5,
+            0.625,
+            0.75,
+            0.875,
+            1.0,
+            1.125,
+            1.25,
+            1.375,
+            1.5,
+            1.625,
+            1.75,
+            1.875,
+            2.0,
+            2.125,
+            2.25,
+            2.375,
+            2.5,
+            2.625,
+            2.75,
+            2.875,
+            3.0,
+            3.125,
+            3.25,
+            3.375,
+            3.5,
+            3.625,
+            3.75,
+            3.875,
+            4.0,
+            4.125,
+            4.25,
+            4.375,
+            4.5,
+            4.625,
+            4.75,
+            4.875,
+            5.0,
+            5.125,
+            5.25,
+            5.375,
+            5.5,
+            5.625,
+            5.75,
+            5.875,
+            6.0,
+            6.125,
+            6.25,
+            6.375,
+            6.5,
+            6.625,
+            6.75,
+            6.875,
+            7.0,
+        ],
+    ],
+}
+
+
+# -----------------------------------------------------------------------------
